Deler KlagebehandlingMapper i to, så en del tar seg av View-objektene, det andre av mappingen til Es
