+++ src/main/kotlin/no/nav/klage/oppgave/api/mapper/KlagebehandlingMapper.kt
@@ -9,8 +9,6 @@ import no.nav.klage.oppgave.clients.pdl.PdlFacade
 import no.nav.klage.oppgave.clients.pdl.Person
 import no.nav.klage.oppgave.clients.saf.rest.ArkivertDokument
 import no.nav.klage.oppgave.domain.elasticsearch.EsKlagebehandling
-import no.nav.klage.oppgave.domain.elasticsearch.EsSaksdokument
-import no.nav.klage.oppgave.domain.elasticsearch.EsVedtak
 import no.nav.klage.oppgave.domain.klage.BrevMottaker
 import no.nav.klage.oppgave.domain.klage.Klagebehandling
 import no.nav.klage.oppgave.domain.klage.PartId
@@ -33,111 +31,6 @@ class KlagebehandlingMapper(
         @Suppress("JAVA_CLASS_ON_COMPANION")
         private val logger = getLogger(javaClass.enclosingClass)
         private val secureLogger = getSecureLogger()
-    }
-
-    fun mapKlagebehandlingOgMottakToEsKlagebehandling(klagebehandling: Klagebehandling): EsKlagebehandling {
-        val klagerFnr = foedselsnummer(klagebehandling.klager.partId)
-        val klagerPersonInfo = klagerFnr?.let { pdlFacade.getPersonInfo(it) }
-
-        val klagerOrgnr = virksomhetsnummer(klagebehandling.klager.partId)
-        val klagerOrgnavn = klagerOrgnr?.let { eregClient.hentOrganisasjon(it)?.navn?.sammensattNavn() }
-
-        val sakenGjelderFnr = foedselsnummer(klagebehandling.sakenGjelder.partId)
-        val sakenGjelderPersonInfo = sakenGjelderFnr?.let { pdlFacade.getPersonInfo(it) }
-
-        val sakenGjelderOrgnr = virksomhetsnummer(klagebehandling.sakenGjelder.partId)
-        val sakenGjelderOrgnavn = sakenGjelderOrgnr?.let { eregClient.hentOrganisasjon(it)?.navn?.sammensattNavn() }
-
-
-        val erFortrolig = sakenGjelderPersonInfo?.harBeskyttelsesbehovFortrolig() ?: false
-        val erStrengtFortrolig = sakenGjelderPersonInfo?.harBeskyttelsesbehovStrengtFortrolig() ?: false
-        val erEgenAnsatt = sakenGjelderFnr?.let { egenAnsattService.erEgenAnsatt(it) } ?: false
-
-        return EsKlagebehandling(
-            id = klagebehandling.id.toString(),
-            versjon = klagebehandling.versjon,
-            klagerFnr = klagerFnr,
-            klagerNavn = klagerPersonInfo?.sammensattNavn,
-            klagerFornavn = klagerPersonInfo?.fornavn,
-            klagerMellomnavn = klagerPersonInfo?.mellomnavn,
-            klagerEtternavn = klagerPersonInfo?.etternavn,
-            klagerOrgnr = klagerOrgnr,
-            klagerOrgnavn = klagerOrgnavn,
-            sakenGjelderFnr = sakenGjelderFnr,
-            sakenGjelderNavn = sakenGjelderPersonInfo?.sammensattNavn,
-            sakenGjelderFornavn = sakenGjelderPersonInfo?.fornavn,
-            sakenGjelderMellomnavn = sakenGjelderPersonInfo?.mellomnavn,
-            sakenGjelderEtternavn = sakenGjelderPersonInfo?.etternavn,
-            sakenGjelderOrgnr = sakenGjelderOrgnr,
-            sakenGjelderOrgnavn = sakenGjelderOrgnavn,
-            tema = klagebehandling.tema.id,
-            type = klagebehandling.type.id,
-            kildeReferanse = klagebehandling.kildeReferanse,
-            sakFagsystem = klagebehandling.sakFagsystem?.id,
-            sakFagsakId = klagebehandling.sakFagsakId,
-            innsendt = klagebehandling.innsendt,
-            mottattFoersteinstans = klagebehandling.mottattFoersteinstans,
-            avsenderSaksbehandleridentFoersteinstans = klagebehandling.avsenderSaksbehandleridentFoersteinstans,
-            avsenderEnhetFoersteinstans = klagebehandling.avsenderEnhetFoersteinstans,
-            mottattKlageinstans = klagebehandling.mottattKlageinstans,
-            tildelt = klagebehandling.tildeling?.tidspunkt,
-            foerstTildelt = klagebehandling.tildelingHistorikk.map { it.tildeling.tidspunkt }.minOrNull(),
-            sistTildelt = klagebehandling.tildeling?.tidspunkt,
-            avsluttet = klagebehandling.avsluttet,
-            avsluttetAvSaksbehandler = klagebehandling.avsluttetAvSaksbehandler,
-            frist = klagebehandling.frist,
-            tildeltSaksbehandlerident = klagebehandling.tildeling?.saksbehandlerident,
-            medunderskriverident = klagebehandling.medunderskriver?.saksbehandlerident,
-            sendtMedunderskriver = klagebehandling.medunderskriver?.tidspunkt,
-            tildeltEnhet = klagebehandling.tildeling?.enhet,
-            hjemler = klagebehandling.hjemler.map { it.id },
-            created = klagebehandling.created,
-            modified = klagebehandling.modified,
-            kilde = klagebehandling.kildesystem.id,
-            kommentarFraFoersteinstans = klagebehandling.kommentarFraFoersteinstans,
-            internVurdering = klagebehandling.kvalitetsvurdering?.internVurdering,
-            vedtak = klagebehandling.vedtak.map { vedtak ->
-                EsVedtak(
-                    utfall = vedtak.utfall?.id,
-                    grunn = vedtak.grunn?.id,
-                    hjemler = vedtak.hjemler.map { hjemmel -> hjemmel.id },
-                    brevmottakerFnr = vedtak.brevmottakere.filter { it.partId.type == PartIdType.PERSON }
-                        .map { it.partId.value },
-                    brevmottakerOrgnr = vedtak.brevmottakere.filter { it.partId.type == PartIdType.VIRKSOMHET }
-                        .map { it.partId.value },
-                    journalpostId = vedtak.journalpostId,
-                    created = vedtak.created,
-                    modified = vedtak.modified,
-                    ferdigstiltIJoark = vedtak.ferdigstiltIJoark
-                )
-            },
-            saksdokumenter = klagebehandling.saksdokumenter.map { EsSaksdokument(it.journalpostId, it.dokumentInfoId) },
-            saksdokumenterJournalpostId = klagebehandling.saksdokumenter.map { it.journalpostId },
-            saksdokumenterJournalpostIdOgDokumentInfoId = klagebehandling.saksdokumenter.map {
-                it.journalpostId + it.dokumentInfoId
-            },
-            egenAnsatt = erEgenAnsatt,
-            fortrolig = erFortrolig,
-            strengtFortrolig = erStrengtFortrolig,
-            vedtakUtfall = klagebehandling.vedtak.firstOrNull()?.utfall?.id,
-            vedtakGrunn = klagebehandling.vedtak.firstOrNull()?.grunn?.id,
-            vedtakHjemler = klagebehandling.vedtak.firstOrNull()?.hjemler?.map { hjemmel -> hjemmel.id } ?: emptyList(),
-            vedtakBrevmottakerFnr = klagebehandling.vedtak.firstOrNull()?.brevmottakere?.filter { it.partId.type == PartIdType.PERSON }
-                ?.map { it.partId.value } ?: emptyList(),
-            vedtakBrevmottakerOrgnr = klagebehandling.vedtak.firstOrNull()?.brevmottakere?.filter { it.partId.type == PartIdType.VIRKSOMHET }
-                ?.map { it.partId.value } ?: emptyList(),
-            vedtakJournalpostId = klagebehandling.vedtak.firstOrNull()?.journalpostId,
-            vedtakCreated = klagebehandling.vedtak.firstOrNull()?.created,
-            vedtakModified = klagebehandling.vedtak.firstOrNull()?.modified,
-            vedtakFerdigstiltIJoark = klagebehandling.vedtak.firstOrNull()?.ferdigstiltIJoark,
-            temaNavn = klagebehandling.tema.name,
-            typeNavn = klagebehandling.type.name,
-            hjemlerNavn = klagebehandling.hjemler.map { it.name },
-            vedtakUtfallNavn = klagebehandling.vedtak.firstOrNull()?.utfall?.name,
-            vedtakGrunnNavn = klagebehandling.vedtak.firstOrNull()?.grunn?.name,
-            sakFagsystemNavn = klagebehandling.sakFagsystem?.name,
-            status = EsKlagebehandling.Status.valueOf(klagebehandling.getStatus().name)
-        )
     }
 
     fun mapEsKlagebehandlingerToListView(
